+++ tests/integ/test_rl.py
@@ -27,8 +27,8 @@ CPU_INSTANCE = 'ml.m4.xlarge'
 
 @pytest.mark.canary_quick
 @pytest.mark.skipif(PYTHON_VERSION != 'py3', reason="RL images supports only Python 3.")
-def test_coach_mxnet(sagemaker_session, rl_coach_full_version):
-    estimator = _test_coach(sagemaker_session, RLFramework.MXNET, rl_coach_full_version)
+def test_coach_mxnet(sagemaker_session, rl_coach_mxnet_full_version):
+    estimator = _test_coach(sagemaker_session, RLFramework.MXNET, rl_coach_mxnet_full_version)
 
     with timeout(minutes=15):
         estimator.fit(wait='False')
@@ -50,8 +50,8 @@ def test_coach_mxnet(sagemaker_session, rl_coach_full_version):
 
 
 @pytest.mark.skipif(PYTHON_VERSION != 'py3', reason="RL images supports only Python 3.")
-def test_coach_tf(sagemaker_session, rl_coach_full_version):
-    estimator = _test_coach(sagemaker_session, RLFramework.TENSORFLOW, rl_coach_full_version)
+def test_coach_tf(sagemaker_session, rl_coach_tf_full_version):
+    estimator = _test_coach(sagemaker_session, RLFramework.TENSORFLOW, rl_coach_tf_full_version)
 
     with timeout(minutes=15):
         estimator.fit()